session_service/core/urls.py
from django.urls import include, path
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView


urlpatterns = [
    path("api/learning-session/", include("SeanceApprentissage.urls")),
]

