session_service/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.2
requests>=2.28
PyPDF2>=3.0.0
pymongo>=4.3

